Move SaveWorld to WorldController
