+++ Assets/Scripts/UI/DialogBox/FileSaveLoad/DialogBoxSaveGame.cs
@@ -96,7 +96,7 @@ public class DialogBoxSaveGame : DialogBoxLoadSaveGame
             // Skip a frame so that user will see pop-up
             yield return null;
 
-            Thread t = SaveWorld(filePath);
+            Thread t = WorldController.Instance.SaveWorld(filePath);
 
             // Wait for data to be saved to HDD.
             while (t.IsAlive)
@@ -116,46 +116,6 @@ public class DialogBoxSaveGame : DialogBoxLoadSaveGame
                 yield return null;
             }
         }
-    }
-
-    /// <summary>
-    /// Serializes current Instance of the World and starts a thread
-    /// that actually saves serialized world to HDD.
-    /// </summary>
-    /// <param name="filePath">Where to save (Full path).</param>
-    /// <returns>Returns the thread that is currently saving data to HDD.</returns>
-    public Thread SaveWorld(string filePath)
-    {
-        // This function gets called when the user confirms a filename
-        // from the save dialog box.
-
-        // Get the file name from the save file dialog box.
-        Debug.ULogChannel("DialogBoxSaveGame", "SaveWorld button was clicked.");
-
-        XmlSerializer serializer = new XmlSerializer(typeof(World));
-        TextWriter writer = new StringWriter();
-        serializer.Serialize(writer, WorldController.Instance.World);
-        writer.Close();
-
-        // UberLogger doesn't handle multi-line messages well.
-        // Debug.Log(writer.ToString());
-
-        // Make sure the save folder exists.
-        if (Directory.Exists(WorldController.Instance.FileSaveBasePath()) == false)
-        {
-            // NOTE: This can throw an exception if we can't create the folder,
-            // but why would this ever happen? We should, by definition, have the ability
-            // to write to our persistent data folder unless something is REALLY broken
-            // with the computer/device we're running on.
-            Directory.CreateDirectory(WorldController.Instance.FileSaveBasePath());
-        }
-
-        // Launch saving operation in a separate thread.
-        // This reduces lag while saving by a little bit.
-        Thread t = new Thread(new ThreadStart(delegate { SaveWorldToHdd(filePath, writer); }));
-        t.Start();
-
-        return t;
     }
 
     /// <summary>
